+++ src/screens/DepositScreen/DepositSuccess/DepositSuccessScreen.styles.ts
